Move more architecture-specific code into a separate folder
