+++ src/arch/x86_64/cpu.h
